config: base64 encoded channels
Most decoders for configuration file formats supported by viper force
UTF-8 content. In both json and toml bytes that are not valid UTF-8 get
turned in to U+FFFD unicode symbols. This prevents interfacing with
channels that predate the world of uniform UTF-8 usage.

To workaround this issue, this change adds a new configuration key for
representing channel names as base64. This key excludes the use of
"Channel" key on the same type.

In addition, a few cases where the channel name was assumed to be all
valid UTF-8 are removed. While RFC2812 agrees with treating channel
names without case sensitivity, it is hard to force such behaviour in
practice due to the wild west of charsets used in channel names.

Work remains to figure out if backwards compatibility for existing
users is required.